+++ nanorq-stable/oblas/gf2.h
@@ -30,13 +30,13 @@ gf2mat *gf2mat_copy(gf2mat *gf2);
 void gf2mat_axpy(gf2mat *gf2, int i, uint8_t *dst, int beta);
 void gf2mat_fill(gf2mat *gf2, int i, uint8_t *dst);
 
-int gf2mat_get(gf2mat *gf2, int i, int j);
-void gf2mat_set(gf2mat *gf2, int i, int j, uint8_t b);
+int gf2mat_get(gf2mat *gf2, size_t i, size_t j);
+void gf2mat_set(gf2mat *gf2, size_t i, size_t j, uint8_t b);
 
 void gf2mat_xor(gf2mat *a, gf2mat *b, int i, int j);
 void gf2mat_and(gf2mat *a, gf2mat *b, int i, int j);
 
-int gf2mat_nnz(gf2mat *gf2, int i, int s, int e);
+int gf2mat_nnz(gf2mat *gf2, size_t i, size_t s, size_t e);
 
 void gf2mat_swaprow(gf2mat *gf2, int i, int j);
 void gf2mat_swapcol(gf2mat *gf2, int i, int j);
