+++ nanorq-stable/oblas/octmat.c
@@ -35,13 +35,13 @@ void om_destroy(octmat *v) {
 void om_print(octmat m, FILE *stream) {
   fprintf(stream, "dense [%ux%u]\n", (unsigned)m.rows, (unsigned)m.cols);
   fprintf(stream, "|     ");
-  for (int j = 0; j < m.cols; j++) {
+  for (size_t j = 0; j < m.cols; j++) {
     fprintf(stream, "| %03d ", j);
   }
   fprintf(stream, "|\n");
-  for (int i = 0; i < m.rows; i++) {
+  for (size_t i = 0; i < m.rows; i++) {
     fprintf(stream, "| %03d | %3d ", i, om_A(m, i, 0));
-    for (int j = 1; j < m.cols; j++) {
+    for (size_t j = 1; j < m.cols; j++) {
       fprintf(stream, "| %3d ", om_A(m, i, j));
     }
     fprintf(stream, "|\n");
